extract context from producer
The producer is automatically adding a trace context to the SQS message
attributes. Read these and use this as the trace ID so that spans in the
consumer are linked to those in the producer.

This gives correct traces in datadog but the flame graphs are very
sparse i.e. there is execution in milliseconds on each side of the queue
and a very large gap of ~10 seconds where the messages are in the queue.